:recycle: Refactor: Request DTO 클래스 record 적용
